Add Transform module and related types/functions
The Transform module and its associated types and functions have been
added. This includes the Transform, TransformFunc, Stride,
TransformFactoryResult, and Transform2FactoryResult structs, as well as
the TransformFn, Transform2Fn, TransformFactory, and Transform2Factory
function types. These changes enable the implementation of
transformations in the library.
